Corrección función char_to_complex
Paso de char a int en la función.

+++ WavReader_Fixed_Sin_Menu.c
@@ -128,26 +128,26 @@ void char_to_complex(char* buff_char, Complex* buff_complex, int len){
         double puente_im[64];
         Complex Aux[64];
 
-        puente_re[0] = fix_to_flo(buff_char[0],FIXED);
-        puente_im[0] = fix_to_flo(buff_char[1],FIXED);
+        puente_re[0] = fix_to_flo((int)buff_char[0],FIXED);
+        puente_im[0] = fix_to_flo((int)buff_char[1],FIXED);
         Aux[0].re = (int)buff_char[0];
         Aux[0].im = (int)buff_char[1];
 
-        puente_re[32] = fix_to_flo(buff_char[64],FIXED);
-        puente_im[32] = fix_to_flo(buff_char[65],FIXED);
+        puente_re[32] = fix_to_flo((int)buff_char[64],FIXED);
+        puente_im[32] = fix_to_flo((int)buff_char[65],FIXED);
         Aux[32].re = buff_char[64];
         Aux[32].im = buff_char[65];
 
         for (int i = 2; i < 64; i = i+2){
-            puente_re[i/2] = fix_to_flo(buff_char[i],FIXED);
-            puente_im[i/2] = fix_to_flo(buff_char[i + 1],FIXED);
-            Aux[i/2].re = buff_char[i];
-            Aux[i/2].im = buff_char[i+1];
+            puente_re[i/2] = fix_to_flo((int)buff_char[i],FIXED);
+            puente_im[i/2] = fix_to_flo((int)buff_char[i + 1],FIXED);
+            Aux[i/2].re = (int)buff_char[i];
+            Aux[i/2].im = (int)buff_char[i+1];
 
-            puente_re[64-i/2] =   fix_to_flo(buff_char[i],FIXED);
-            puente_im[64-i/2] = - fix_to_flo(buff_char[i + 1],FIXED);
-            Aux[64 - i/2].re = buff_char[i];
-            Aux[64 - i/2].im = buff_char[i+1];
+            puente_re[64-i/2] =   fix_to_flo((int)buff_char[i],FIXED);
+            puente_im[64-i/2] = - fix_to_flo((int)buff_char[i + 1],FIXED);
+            Aux[64 - i/2].re = (int)buff_char[i];
+            Aux[64 - i/2].im = (int)buff_char[i+1];
         }
 
         for (int i = 0; i < 64; i++){
